+++ packages/datepicker-suggest-react/src/date-picker-suggest.tsx
@@ -10,7 +10,6 @@ import {
   ComboboxInput,
   ComboboxOption,
   ComboboxOptions,
-  Label,
 } from "@headlessui/react";
 
 const DropdownIcon = () => (
@@ -31,7 +30,9 @@ const DropdownIcon = () => (
 type DatePickerSuggestProps = {
   value?: Date;
   onChange?: (date: Date | undefined) => void;
+  panelClassName?: string;
   suggestionRenderer?: (dateOption: { date: Date; label: string }) => ReactNode;
+  displayValue?: (dateOption: { date: Date; label: string }) => string;
 };
 
 const defaultSuggestionRenderer = (dateSuggestion: DateSuggestion) => (
@@ -39,16 +40,24 @@ const defaultSuggestionRenderer = (dateSuggestion: DateSuggestion) => (
     <span className="truncate group-data-[selected]:font-semibold">
       {dateSuggestion.label}
     </span>
-    <span className="ml-2 truncate text-gray-400 justify-end">
+    <span className="truncate text-gray-400 ml-auto">
       {format(dateSuggestion.date, "dd MMM yyyy HH:mm")}
     </span>
   </>
 );
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const defaultDisplayValue = (dateSuggestion: DateSuggestion) =>
+  dateSuggestion.label;
+
 export const DatePickerSuggest = (props: DatePickerSuggestProps) => {
   const suggestionRenderer =
     props.suggestionRenderer ?? defaultSuggestionRenderer;
+  const displayValue = props.displayValue ?? defaultDisplayValue;
+  let panelClassName = props.panelClassName ?? "";
+  if (!panelClassName.includes("w-")) {
+    panelClassName += " w-[var(--input-width)]";
+  }
+
   const suggest = new SuggestionEngine();
 
   const [inputValue, setInputValue] = useState("");
@@ -59,7 +68,7 @@ export const DatePickerSuggest = (props: DatePickerSuggestProps) => {
   console.log("bog: props.value", props.value);
 
   useEffect(() => {
-    if (props.value != selectedDate?.date) {
+    if (props.value?.getTime() != selectedDate?.date?.getTime()) {
       const propsSelectedDate = props.value
         ? {
             date: props.value,
@@ -92,21 +101,15 @@ export const DatePickerSuggest = (props: DatePickerSuggestProps) => {
         }}
         onClose={() => setInputValue("")}
       >
-        <Label className="block text-sm font-medium leading-6 text-gray-100">
-          Date
-        </Label>
         <div className="relative">
           <ComboboxInput
             autoComplete="off"
-            className="w-full rounded-lg border-none bg-white/5 py-1.5 pr-8 pl-3 text-sm/6 text-white focus:outline-none data-[focus]:outline-2 data-[focus]:-outline-offset-2 data-[focus]:outline-white/25"
+            className="w-full rounded-lg border-none bg-white/5 py-1.5 pr-8 pl-3 text-sm/6 focus:outline-none data-[focus]:outline-2 data-[focus]:-outline-offset-2 data-[focus]:outline-white/25"
             onChange={handleInputChange}
             onBlur={() => setInputValue("")}
             displayValue={(dateSuggestion) =>
               !!dateSuggestion
-                ? format(
-                    (dateSuggestion as DateSuggestion).date,
-                    "dd MMM yyyy HH:mm"
-                  )
+                ? displayValue(dateSuggestion as DateSuggestion)
                 : ""
             }
           />
@@ -123,7 +126,7 @@ export const DatePickerSuggest = (props: DatePickerSuggestProps) => {
             <ComboboxOptions
               anchor="bottom"
               transition
-              className="w-[var(--input-width)] rounded-xl border border-white/5 bg-white/5 p-1 [--anchor-gap:var(--spacing-1)] empty:invisible transition duration-100 ease-in data-[leave]:data-[closed]:opacity-0"
+              className={`${panelClassName} rounded-xl border dark:text-white dark:bg-black text-black bg-white border-white/5 bg-white/5 p-1 [--anchor-gap:var(--spacing-1)] empty:invisible transition duration-100 ease-in data-[leave]:data-[closed]:opacity-0`}
             >
               {result.map((suggestionResult) => (
                 <ComboboxOption
@@ -131,15 +134,9 @@ export const DatePickerSuggest = (props: DatePickerSuggestProps) => {
                   value={suggestionResult}
                   className="group flex cursor-default items-center gap-2 rounded-lg py-1.5 px-3 select-none data-[focus]:bg-white/10"
                 >
-                  <div className="flex text-sm/6 text-white">
+                  <div className="flex w-full text-sm/6 text-white">
                     {suggestionRenderer(suggestionResult)}
                   </div>
-
-                  {/* <span className="invisible size-4 fill-white group-data-[selected]:visible">
-                    <span className="h-5 w-5" aria-hidden="true">
-                      <CheckIcon />
-                    </span>
-                  </span> */}
                 </ComboboxOption>
               ))}
             </ComboboxOptions>