+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @datepicker-suggest/react@0.0.4
+  - @datepicker-suggest/core@0.0.4
+
 ## 0.1.2
 
 ### Patch Changes