+++ packages/datepicker-suggest-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @datepicker-suggest/core
 
+## 0.0.4
+
+### Patch Changes
+
+- fixes and new props exposed
+
 ## 0.0.3
 
 ### Patch Changes