+++ packages/datepicker-suggest-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @datepicker-suggest/react
 
+## 0.0.4
+
+### Patch Changes
+
+- fixes and new props exposed
+- Updated dependencies
+  - @datepicker-suggest/core@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes